+++ cmd/harp/internal/cmd/transform_encode.go
@@ -18,8 +18,6 @@
 package cmd
 
 import (
-	"encoding/base64"
-	"encoding/hex"
 	"io"
 
 	"github.com/spf13/cobra"
@@ -27,6 +25,7 @@ import (
 
 	"github.com/elastic/harp/pkg/sdk/cmdutil"
 	"github.com/elastic/harp/pkg/sdk/log"
+	"github.com/elastic/harp/pkg/sdk/value/encoding"
 )
 
 // -----------------------------------------------------------------------------
@@ -40,9 +39,32 @@ type transformEncodeParams struct {
 var transformEncodeCmd = func() *cobra.Command {
 	params := &transformEncodeParams{}
 
+	longDesc := cmdutil.LongDesc(`
+	Encode the given input stream using the selected encoding strategy.
+
+	Supported codecs:
+	  * identity - returns the unmodified input
+	  * hex/base16 - returns the hexadecimal encoded input
+	  * base32 - returns the Base32 encoded input
+	  * base32hex - returns the Base32 with extended alphabet encoded input
+	  * base64 - returns the Base64 encoded input
+	  * base64raw - returns the Base64 encoded input without "=" padding
+	  * base64url - returns the Base64 encoded input using URL safe characters
+	  * base64urlraw - returns the Base64 encoded input using URL safe characters without "=" padding
+	  * base85 - returns the Base85 encoded input`)
+
+	examples := cmdutil.Examples(`
+		# Encode stdin to base64
+		echo "test" | harp transform encode --encoding base64
+
+		# Encode a file using base64url
+		harp transform encode --encoding base64url --in test.txt --out encoded.text `)
+
 	cmd := &cobra.Command{
-		Use:   "encode",
-		Short: "Encode given input",
+		Use:     "encode",
+		Short:   "Encode given input",
+		Long:    longDesc,
+		Example: examples,
 		Run: func(cmd *cobra.Command, args []string) {
 			// Initialize logger and context
 			ctx, cancel := cmdutil.Context(cmd.Context(), "harp-transform-encode", conf.Debug.Enable, conf.Instrumentation.Logs.Level)
@@ -60,43 +82,28 @@ var transformEncodeCmd = func() *cobra.Command {
 				log.For(ctx).Fatal("unable to initialize output writer", zap.Error(err))
 			}
 
-			// Drain reader
-			content, err := io.ReadAll(reader)
+			// Write encoded content
+			encoderWriter, err := encoding.NewWriter(writer, params.encoding)
 			if err != nil {
-				log.For(ctx).Fatal("unable to drain input reader", zap.Error(err))
+				log.SafeClose(encoderWriter, "unable to close the encoder writer")
+				log.For(ctx).Fatal("unable to write encoded content", zap.Error(err))
 			}
 
-			var out string
-
-			// Apply transformation
-			switch params.encoding {
-			case "identity":
-				out = string(content)
-			case "hex":
-				out = hex.EncodeToString(content)
-			case "base64":
-				out = base64.StdEncoding.EncodeToString(content)
-			case "base64raw":
-				out = base64.RawStdEncoding.EncodeToString(content)
-			case "base64url":
-				out = base64.URLEncoding.EncodeToString(content)
-			case "base64urlraw":
-				out = base64.RawURLEncoding.EncodeToString(content)
-			default:
-				log.For(ctx).Fatal("unhandled encoding strategy", zap.String("encoding", params.encoding))
+			// Process input as a stream.
+			if _, err := io.Copy(encoderWriter, reader); err != nil {
+				log.SafeClose(encoderWriter, "unable to close the encoder writer")
+				log.For(ctx).Fatal("unable to process input", zap.Error(err))
 			}
 
-			// Dump as output
-			if _, err = writer.Write([]byte(out)); err != nil {
-				log.For(ctx).Fatal("unable to write result to writer", zap.Error(err))
-			}
+			// Close the writer
+			log.SafeClose(encoderWriter, "unable to close the encoder writer")
 		},
 	}
 
 	// Parameters
 	cmd.Flags().StringVar(&params.inputPath, "in", "-", "Input path ('-' for stdin or filename)")
-	cmd.Flags().StringVar(&params.outputPath, "out", "-", "Output path ('-' for stdin or filename)")
-	cmd.Flags().StringVar(&params.encoding, "encoding", "identity", "Encoding strategy (hex, base64, base64raw, base64url, base64urlraw)")
+	cmd.Flags().StringVar(&params.outputPath, "out", "-", "Output path ('-' for stdout or filename)")
+	cmd.Flags().StringVar(&params.encoding, "encoding", "identity", "Encoding strategy")
 
 	return cmd
 }