+++ cmd/harp/internal/cmd/transform_encrypt.go
@@ -18,26 +18,30 @@
 package cmd
 
 import (
-	"encoding/base64"
 	"io"
+	"strings"
 
 	"github.com/spf13/cobra"
 	"go.uber.org/zap"
 
 	"github.com/elastic/harp/pkg/sdk/cmdutil"
 	"github.com/elastic/harp/pkg/sdk/log"
+	"github.com/elastic/harp/pkg/sdk/value/encoding"
 	"github.com/elastic/harp/pkg/sdk/value/encryption"
 )
 
 // -----------------------------------------------------------------------------
 
+type transformEncryptParams struct {
+	inputPath              string
+	outputPath             string
+	keyRaw                 string
+	additionalData         string
+	additionalDataEncoding string
+}
+
 var transformEncryptCmd = func() *cobra.Command {
-	var (
-		inputPath      string
-		outputPath     string
-		keyRaw         string
-		additionalData string
-	)
+	params := &transformEncryptParams{}
 
 	cmd := &cobra.Command{
 		Use:     "encrypt",
@@ -49,7 +53,7 @@ var transformEncryptCmd = func() *cobra.Command {
 			defer cancel()
 
 			// Resolve tranformer
-			t, err := encryption.FromKey(keyRaw)
+			t, err := encryption.FromKey(params.keyRaw)
 			if err != nil {
 				log.For(ctx).Fatal("unable to initialize a transformer form key", zap.Error(err))
 			}
@@ -58,13 +62,13 @@ var transformEncryptCmd = func() *cobra.Command {
 			}
 
 			// Read input
-			reader, err := cmdutil.Reader(inputPath)
+			reader, err := cmdutil.Reader(params.inputPath)
 			if err != nil {
 				log.For(ctx).Fatal("unable to initialize input reader", zap.Error(err))
 			}
 
 			// Read input
-			writer, err := cmdutil.Writer(inputPath)
+			writer, err := cmdutil.Writer(params.outputPath)
 			if err != nil {
 				log.For(ctx).Fatal("unable to initialize output writer", zap.Error(err))
 			}
@@ -76,10 +80,15 @@ var transformEncryptCmd = func() *cobra.Command {
 			}
 
 			// Decode AAD if any
-			if additionalData != "" {
-				aad, errDecode := base64.StdEncoding.DecodeString(additionalData)
+			if params.additionalData != "" {
+				encoderReader, errDecode := encoding.NewReader(strings.NewReader(params.additionalData), params.additionalDataEncoding)
 				if errDecode != nil {
 					log.For(ctx).Fatal("unable to decode additional data", zap.Error(errDecode))
+				}
+
+				aad, err := io.ReadAll(encoderReader)
+				if err != nil {
+					log.For(ctx).Fatal("unable to read additional data", zap.Error(err))
 				}
 
 				// Set additional data
@@ -99,12 +108,13 @@ var transformEncryptCmd = func() *cobra.Command {
 	}
 
 	// Parameters
-	cmd.Flags().StringVar(&keyRaw, "key", "", "Transformer key")
+	cmd.Flags().StringVar(&params.keyRaw, "key", "", "Transformer key")
 	log.CheckErr("unable to mark 'key' flag as required.", cmd.MarkFlagRequired("key"))
 
-	cmd.Flags().StringVar(&inputPath, "in", "-", "Input path ('-' for stdin or filename)")
-	cmd.Flags().StringVar(&outputPath, "out", "-", "Output path ('-' for stdin or filename)")
-	cmd.Flags().StringVar(&additionalData, "aad", "", "Standard BASE64 encoded additional data for AEAD encryption")
+	cmd.Flags().StringVar(&params.inputPath, "in", "-", "Input path ('-' for stdin or filename)")
+	cmd.Flags().StringVar(&params.outputPath, "out", "-", "Output path ('-' for stdout or filename)")
+	cmd.Flags().StringVar(&params.additionalData, "aad", "", "Additional data for AEAD encryption")
+	cmd.Flags().StringVar(&params.additionalDataEncoding, "aad-encoding", "base64", "Additional data encoding strategy")
 
 	return cmd
 }