+++ cmd/harp/internal/cmd/transform_verify.go
@@ -91,7 +91,7 @@ var transformVerifyCmd = func() *cobra.Command {
 	log.CheckErr("unable to mark 'key' flag as required.", cmd.MarkFlagRequired("key"))
 
 	cmd.Flags().StringVar(&inputPath, "in", "-", "Input path ('-' for stdin or filename)")
-	cmd.Flags().StringVar(&outputPath, "out", "-", "Output path ('-' for stdin or filename)")
+	cmd.Flags().StringVar(&outputPath, "out", "-", "Output path ('-' for stdout or filename)")
 
 	return cmd
 }